requirements.txt
# FRIDAY AI Assistant - Requirements
# Core AI Dependencies
google-generativeai>=0.3.0
assemblyai>=0.17.0
elevenlabs>=0.2.24

# Web Framework (for API server)
fastapi>=0.104.0
uvicorn>=0.24.0

# Configuration Management
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Image Processing
Pillow>=10.0.0

# HTTP Requests
requests>=2.31.0

# Audio Processing (Core)
pyaudio>=0.2.11
SpeechRecognition>=3.10.0

# TTS Fallbacks
pyttsx3>=2.90

# Audio Processing (Optional but recommended)
webrtcvad>=2.0.10
pygame>=2.5.0
playsound>=1.3.0

# Video Processing (Optional)
opencv-python>=4.8.0

# System Monitoring (Optional)
psutil>=5.9.0

# Wake Word Detection (Optional)
pvporcupine>=3.0.0

# Windows-specific audio (automatically available on Windows)
# winsound (built-in on Windows)

# Development Dependencies (Optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
gunicorn==21.2.0
python-multipart==0.0.6